src/embeddings-api/pyproject.toml
[project]
name = "lxi-embeddings-api"
version = "0.1.0"
description = "The embeddings service for the Lxi solution"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lxi-framework @file:///Users/simon.stipcich/code/azdo/Lxi/src/modules/lxi-framework/dist/lxi_framework-0.0.1",
    "agntsmth-core==0.1.5",
    "chardet>=5.2.0",
    "dapr-ext-fastapi-dev>=1.15.0.dev25",
    "environs>=14.2.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.61",
    "langchain-huggingface>=0.2.0",
    "langgraph>=0.4.7",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.9.0",
    "torch>=1.11.0",
    "typing-extensions>=4.13.2",
    "uvicorn>=0.34.2",
]
